server/train/requirements.txt
functions-framework==3.*
pandas
nltk
git+https://github.com/UKPLab/sentence-transformers
np
google-cloud-storage
transformers